#8 Refactor code
